Validate UUIDs and standardize error handling across services
Added UUID validation checks in user, follower, and post services to ensure input consistency and avoid invalid IDs. Updated error messages for clarity and accuracy, aligning responses with standardized formats across the application.

+++ node/express-prisma-typescript/src/domains/user/service/user.service.impl.ts
@@ -1,15 +1,17 @@
-import { NotFoundException } from '@utils/errors'
+import { NotFoundException, ValidationException } from '@utils/errors'
 import { OffsetPagination } from 'types'
 import { UserDTO, UserViewDTO } from '../dto'
 import { UserRepository } from '../repository'
 import { UserService } from './user.service'
 import { generatePresignedUrl, getPublicUrl } from '@utils/s3-utils'
 import { v4 as uuidv4 } from 'uuid'
+import { isUuid } from 'uuidv4'
 
 export class UserServiceImpl implements UserService {
   constructor (private readonly repository: UserRepository) {}
 
   async getUser (userId: any): Promise<UserViewDTO> {
+    if (!isUuid(userId)) throw new ValidationException([{ message: 'INVALID_UUID' }])
     const user = await this.repository.getById(userId)
     if (!user) throw new NotFoundException('user')
     return user
@@ -21,10 +23,13 @@ export class UserServiceImpl implements UserService {
   }
 
   async deleteUser (userId: any): Promise<void> {
+    if (!isUuid(userId)) throw new ValidationException([{ message: 'INVALID_UUID' }])
+    if (await this.repository.getById(userId) === null) throw new NotFoundException('user')
     await this.repository.delete(userId)
   }
 
   async isPrivate (userId: string): Promise<boolean> {
+    if (!isUuid(userId)) throw new ValidationException([{ message: 'INVALID_UUID' }])
     const user = await this.repository.getById(userId)
     if (!user) throw new NotFoundException('user')
     return await this.repository.isPrivate(userId)
@@ -48,6 +53,7 @@ export class UserServiceImpl implements UserService {
   }
 
   async updateUserPrivacy (userId: string, isPrivate: boolean): Promise<UserViewDTO> {
+    if (!isUuid(userId)) throw new ValidationException([{ message: 'INVALID_UUID' }])
     return await this.repository.updatePrivacy(userId, isPrivate)
   }
 }
